# add parallelizer and benchmark test in estimator for performance