@Garrybest 2022-03-30
Before:

```
goos: linux
goarch: amd64
pkg: github.com/karmada-io/karmada/pkg/estimator/server
cpu: Intel(R) Xeon(R) Gold 6133 CPU @ 2.50GHz
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/500_nodes_and_10,000_pods_without_affinity_and_tolerations-8                  166           7034555 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_without_affinity_and_tolerations-8                 13          85556061 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_taint_and_tolerations-8                       13          87564978 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_node_affinity_and_tolerations-8               13          90424776 ns/op
PASS
ok      github.com/karmada-io/karmada/pkg/estimator/server      19.239s
```

After：

```
goos: linux
goarch: amd64
pkg: github.com/karmada-io/karmada/pkg/estimator/server
cpu: Intel(R) Xeon(R) Gold 6133 CPU @ 2.50GHz
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/500_nodes_and_10,000_pods_without_affinity_and_tolerations-8                  684           1690992 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_without_affinity_and_tolerations-8                 64          18327287 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_taint_and_tolerations-8                       74          17490168 ns/op
BenchmarkAccurateSchedulerEstimatorServer_MaxAvailableReplicas/5000_nodes_and_100,000_pods_with_node_affinity_and_tolerations-8               74          18474715 ns/op
PASS
ok      github.com/karmada-io/karmada/pkg/estimator/server      18.318s
```