@Garrybest 2022-03-30
General speedup ratio: 4-5 times

In a cluster with 5000 nodes and 100,000 pods, the cost is less than 20ms, which does not count on a gRPC transmission cost.

My machine is a cvm used for development, maybe faster when in a production environment.

/cc @RainbowMango @XiShanYongYe-Chang @mrlihanbo @huone1