@Garrybest 2022-03-31
> the two commits are repeat

The first one is a benchmark test, the other one is an optimization. By adding benchmark first, we could test the performance before.